Create Requirements
INTRODUCTION
The data analyst’s job does not end once the data has been analyzed. A vital skill for data analysts is to represent and report the data to stakeholders. In this task, you will demonstrate your ability to identify actionable insights from data and communicate them using effective storytelling methods.



This task consists of three parts:



In part 1, you will use SQL and a business intelligence tool to create data dashboards to support executive decision-making. Your dashboards should enable leaders to answer a specific research question on an organizational need. It is recommended, but not required, that the business intelligence tool be Tableau.



Your dashboard will integrate data from two sources:



1.  Provided data set: You will select one of the data sets and associated data dictionaries from the web link "D211 Data Sets and Associated Data Dictionaries" located on Labs on Demand.



Note: It is recommended, but not required, that you choose the same data set that you used for your performance assessments in previous courses. Review the data dictionary that is relevant to your chosen data set for additional details about the audience.


2.  Additional data set: You will identify an external, public data set as an additional data set for your dashboard. This data set should contain data and variables that complement the data set you chose from the provided options and should enhance the insights you can gain from the provided data set. Recommended sources include U.S. census data, Kaggle, or other public data repositories. You are responsible for ensuring that you have the rights to use the data set.



In part 2, you will give a simulated presentation to a panel of peers. You will describe the use of SQL in preparing the data from the two data sets for the analysis and then demonstrate the functionality of the SQL scripts and other codes that supported the creation of the dashboards.



In part 3, you will write a reflection paper to outline the organizational need and summarize the findings from the analysis.



Refer to the data dictionary for your chosen provided data set for additional background information about the audience of your data representations.



REQUIREMENTS
Your submission must be your original work. No more than a combined total of 30% of the submission and no more than a 10% match to any one individual source can be directly quoted or closely paraphrased from sources, even if cited correctly. The similarity report that is provided when you submit your task can be used as a guide.



You must use the rubric to direct the creation of your submission because it provides detailed criteria that will be used to evaluate your work. Each requirement below may be evaluated by more than one rubric aspect. The rubric aspect titles may contain hyperlinks to relevant portions of the course.



Tasks may not be submitted as cloud links, such as links to Google Docs, Google Slides, OneDrive, etc., unless specified in the task requirements. All other submissions must be file types that are uploaded and submitted as attachments (e.g., .csv, .docx, .pdf, .ppt).



To work on this task, you must follow the link "D211 Data Sets and Associated Dictionaries" in the Web Links section to the LoDs PostgreSQL lab environment. In this environment, you will be able to write and test your PostgreSQL code and access the databases to complete this task.



Part 1:  Data Dashboards

A.  Provide a copy of your dashboards that support executive decision-making.

1.  Provide both data sets that serve as the data source for the dashboards.

2.  Provide step-by-step instructions to guide users through the dashboard installation.

3.  Provide clear instructions to help users navigate the dashboards.

4.  Provide a copy of all SQL code and other code supporting the dashboards.


Part 2:  Demonstration

B.  Provide a link to a Panopto multimedia presentation in which you present the dashboards to an audience of data analytics peers. You should do all of the following in your presentation:

1.  Describe the technical environment used to create the dashboards.

2.  Demonstrate the functionality of the dashboards.

3.  Explain the SQL scripts used to support the creation of the dashboards.

4.  Explain how the data streams were prepared to support the analysis.

5.  Describe how data were aligned with other data points.

6.  Demonstrate how the databases were created.

7.  Explain how referential integrity was enforced in the database.


Note: The audiovisual recording should feature you presenting the material on screen (i.e., not in voice-over or embedded video) and should simultaneously capture both you and your multimedia presentation.


Note: For instructions on how to access and use Panopto, use the "Panopto How-To Videos" web link provided below. To access Panopto's website, navigate to the web link titled "Panopto Access," and then choose to log in using the “WGU” option. If prompted, log in using your WGU student portal credentials, and then it will forward you to Panopto’s website.


To submit your recording, upload it to the Panopto drop box titled “Master of Science, Advanced Data Acquisition SLM1 | D211 (Student Creators) [assignments]." Once the recording has been uploaded and processed in Panopto's system, retrieve the URL of the recording from Panopto and copy and paste it into the Links option. Upload the remaining task requirements using the Attachments option.


Part 3:  Report

C.  Write a report to outline the data exploration, use of advanced SQL operations, and the analysis of the data. Do the following as part of your report:

1.  Explain how the purpose and function of your dashboard aligns with the needs outlined in the data dictionary associated with your chosen data set.

2.  Justify the selection of the business intelligence tool you used.

3.  Explain the steps used to clean and prepare the data for the analysis.

4.  Summarize the steps used to create the dashboards.

5.  Discuss the results of your data analysis and how it supports executive decision-making.

6.  Discuss the limitation(s) of your data analysis.


D.  Record the web sources used to acquire data or segments of third-party code used to support the application. Ensure the web sources cited are reliable.


E.  Acknowledge sources, using in-text citations and references, for content that is quoted, paraphrased, or summarized.


F.  Demonstrate professional communication in the content and presentation of your submission.